Refactor package hierarchy, shift Bob and Ui to bob.cli
